ENH: Make the remaining integer conversions explicit
The -Wsign-conversion findings that are not allocation, copy or I/O
lengths: file offsets and positions, extension bookkeeping, and a few
conversions that were wrong rather than merely implicit.

  File offsets.  znztell() returns a signed znz_off_t stored in size_t
  locals; the znzseek() arguments are cast to znz_off_t so the arithmetic
  is done at the width the function takes.

  fslio's nvox.  The product of seven int dimensions assigned to a size_t
  field, widened per operand so it is computed at 64 bits.

  The ANALYZE orientation byte.  nifti_convert_nhdr2nim() read it with
  `unsigned char c = *((char *)(&nhdr.qform_code));`, taking a signed
  char and widening it into an unsigned one, so any value above 127
  arrived sign-extended.  Reading through unsigned char * gives the byte
  that is actually in the header.

  expat's XML_Parse takes an int length while blen is unsigned; the
  conversion is now at the call rather than hidden in the variable.

Lines that upstream PRs also change are left alone, including the
d3matrix and d4matrix allocations in fslio.c, which PR #22 rewrites.

+++ cifti/afni_xml.c
@@ -223,7 +223,7 @@ afni_xml_list axml_read_file(const char * fname, int read_data)
          if( xd->verb > 1 )
             fprintf(stderr,"-- AXML: truncating fbuffer from %u to %" PRId64  "\n",
                     blen, bshort);
-         blen = (int)bshort;
+         blen = (unsigned)bshort;
       }
       /* a short read means end of input; so does hitting EOF exactly on a
          buffer boundary, and stopping here avoids a final read that can
@@ -306,7 +306,7 @@ afni_xml_list axml_read_buf(const char * buf_in, int64_t bin_len)
         /*--- replace fread with buffer copy ---*/
 
         /* decide how much to copy and copy */
-        if( bin_remain >= bsize ) blen = bsize;
+        if( bin_remain >= bsize ) blen = (unsigned)bsize;
         else                      blen = (unsigned)bin_remain;
 
         if(blen > 0 && blen <= (unsigned)bsize) {
@@ -676,7 +676,7 @@ static int reset_xml_buf(afni_xml_control * xd, char ** buf, int * bsize)
         fprintf(stderr,"++ update buf, %d to %d bytes\n",*bsize,xd->buf_size);
 
     *bsize = xd->buf_size;
-    *buf = (char *)safe_realloc(*buf,(size_t)((*bsize+1) * sizeof(char)));
+    *buf = (char *)safe_realloc(*buf,(size_t)(*bsize+1) * sizeof(char));
     if( ! *buf ) {
         fprintf(stderr,"** failed to alloc %d bytes of xml buf!\n", *bsize);
         *bsize = 0;
@@ -897,7 +897,7 @@ static const char * strip_whitespace(const char * str, int slen)
 
    /* make sure we have local space */
    if( len > blen ) { /* allocate a bigger buffer */
-      buf = (char *)safe_realloc(buf,(size_t)((len+1) * sizeof(char)));
+      buf = (char *)safe_realloc(buf,(size_t)(len+1) * sizeof(char));
       if( !buf ) {
          fprintf(stderr,"** failed to alloc wspace buf of len %d\n", len+1);
          return str;
